Moved autocomplete.h/cpp from win32 to main folders
Also enabled building for linux, provided the compiler supports c++11 or later.
This enables autocomplete use for megaclc in the MEGAchat project.
