+++ contrib/cmake/CMakeLists.txt
@@ -50,13 +50,13 @@ endif(build_64_bit)
 IF(WIN32)
     set(MegaDir "${CMAKE_CURRENT_LIST_DIR}/../..")
     if ("${Mega3rdPartyDir}" STREQUAL "")
-        set(Mega3rdPartyDir "${MegaDir}/../3rdparty")
+        set(Mega3rdPartyDir "${MegaDir}/../3rdParty")
     endif()
 ELSE(WIN32)
     set(MegaDir "${CMAKE_CURRENT_LIST_DIR}/../..")
-    set(Mega3rdPartyDir "${MegaDir}/../3rdParty/vcpkg/installed/x64-linux")
-    include_directories( "${MegaDir}/include" "${Mega3rdPartyDir}/include" )
-    link_directories( "${Mega3rdPartyDir}/lib" )
+    if ("${Mega3rdPartyDir}" STREQUAL "")
+        set (Mega3rdPartyDir "${MegaDir}/../3rdParty/")
+    endif()
 ENDIF(WIN32)
 
 IF(WIN32)
@@ -65,6 +65,10 @@ IF(WIN32)
     else(build_64_bit)
         set(vcpkg_dir "${Mega3rdPartyDir}/vcpkg/installed/x86-windows-static-uncheckediterators")
     endif(build_64_bit)
+ELSE(WIN32)
+    set(vcpkg_dir "${Mega3rdPartyDir}/vcpkg/installed/x64-linux")
+    include_directories( "${MegaDir}/include" "${Mega3rdPartyDir}/include" )
+    link_directories( "${Mega3rdPartyDir}/lib" )
 ENDIF(WIN32)
 
 if (NOT CMAKE_BUILD_TYPE)
@@ -218,7 +222,6 @@ IF(WIN32)
     SET(Mega_PlatformSpecificLibs ws2_32 winhttp Shlwapi Secur32.lib)
 
     SET(Mega_PlatformSpecificFiles ${MegaDir}/src/win32/console.cpp 
-    ${MegaDir}/src/win32/autocomplete.cpp 
     ${MegaDir}/src/win32/consolewaiter.cpp 
     ${MegaDir}/src/win32/fs.cpp 
     $<IF:${USE_CURL},${MegaDir}/src/posix/net.cpp,${MegaDir}/src/win32/net.cpp>
@@ -254,6 +257,7 @@ SET(Mega_GfxFiles ${MegaDir}/src/gfx/external.cpp ${MegaDir}/src/gfx/freeimage.c
 
 add_library(Mega STATIC
             ${MegaDir}/src/attrmap.cpp 
+            ${MegaDir}/src/autocomplete.cpp 
             ${MegaDir}/src/backofftimer.cpp 
             ${MegaDir}/src/base64.cpp 
             ${MegaDir}/src/command.cpp
